Pull out form list parsing
